% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choices_from_relevancy.R
\name{choices_from_relevancy}
\alias{choices_from_relevancy}
\title{Extract choices list from relevancy rule defined}
\usage{
choices_from_relevancy(relevancy_string)
}
\arguments{
\item{relevancy_string}{The relevancy formula for which the list of choices is needed}
}
\value{
The list of choices involved in the relevancy formula
}
\description{
Extract choices list from relevancy rule defined
}
