% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_dt.R
\name{compare_dt}
\alias{compare_dt}
\title{Compare data frames}
\usage{
compare_dt(
  df1,
  df2,
  unique_id_df1 = "KEY",
  unique_id_df2 = "KEY",
  compare_all = TRUE
)
}
\arguments{
\item{df1}{old version of the dataset}

\item{df2}{latest version of the dataset}

\item{unique_id_df1}{unique identifier in df1, defaults to KEY}

\item{unique_id_df2}{unique identifier in df2, defaults to KEY}

\item{compare_all}{logical. TRUE: compare all columns/variables. FALSE: compare only shared columns/variables.}
}
\description{
The \code{compare_dt()} function compares two versions of the same data
}
\examples{
library(atRfunctions)

mtcars_v1 <- mtcars_v2 <- mtcars
mtcars_v1$id <- mtcars_v2$id <- rownames(mtcars)

mtcars_v2$am[mtcars_v2$am == 1] <- 2
mtcars_v2$wt[mtcars_v2$wt > 100] <- 105
mtcars_v2$cyl[mtcars_v2$cyl == 8] <- 7
mtcars_v2 <- mtcars_v2[1:30, ]
mtcars_v2$row_n <- 1:nrow(mtcars_v2)

compare_dt(df1 = mtcars_v1, df2 = mtcars_v2,
          unique_id_df1 = "id",
          unique_id_df2 = "id",
          compare_all = FALSE
)

compare_dt(df1 = mtcars_v1, df2 = mtcars_v2,
          unique_id_df1 = "id",
          unique_id_df2 = "id",
          compare_all = TRUE
)

}
