% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_directory.R
\name{create_directory}
\alias{create_directory}
\title{Checks if the path exists, if not, creates it}
\usage{
create_directory(path)
}
\arguments{
\item{path}{The directory path to create}
}
\description{
This function checks if the given path/directory exists. If not, it will create it
}
