% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions_from_relevancy.R
\name{questions_from_relevancy}
\alias{questions_from_relevancy}
\title{Extract question list from relevancy rule defined}
\usage{
questions_from_relevancy(relevancy_string)
}
\arguments{
\item{relevancy_string}{The relevancy formula for which the list of question is needed}
}
\description{
Extract question list from relevancy rule defined
}
