% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xlsx_sheets.R
\name{read_xlsx_sheets}
\alias{read_xlsx_sheets}
\title{Reading XLSX data as a list}
\usage{
read_xlsx_sheets(path, guess_max = 5e+06, convert_to_na = c("N/A", "-", " "))
}
\arguments{
\item{path}{Path to the XLSX file}

\item{guess_max}{Guess Max ; defaults to 5000000}

\item{convert_to_na}{a list of patterns to be converted to NA ; defaults to c("N/A", "-", " ")}
}
\value{
A list of data set where each item represent XLSX's sheet
}
\description{
Read XLSX as list
}
