% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_to_data_merge.R
\name{reshape_to_datamerge}
\alias{reshape_to_datamerge}
\title{Reshape the analysis result file into Data Merge format for Adobe InDesign}
\usage{
reshape_to_datamerge(analysis_result)
}
\arguments{
\item{analysis_result}{Data Analysis Result Dataset}
}
\value{
dataframe
}
\description{
Reshape the analysis result file into Data Merge format for Adobe InDesign
}
