% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_repeat_sheet_names_to_questions.R
\name{add_repeat_sheet_names_to_questions}
\alias{add_repeat_sheet_names_to_questions}
\title{Identify and set the sheets names for each question in Data collection tool}
\usage{
add_repeat_sheet_names_to_questions(kobo_tool)
}
\arguments{
\item{kobo_tool}{the kobo tool for which the question's sheet needed}
}
\description{
Identify and set the sheets names for each question in Data collection tool
}
