% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_log.R
\name{apply_log}
\alias{apply_log}
\title{Apply data cleaning/change log}
\usage{
library(atRfunctions)
apply_log(data, log,
          data_KEY = "KEY",
          log_columns = c(question = "question",
                          old_value = "old_value",
                          new_value = "new_value",
                          KEY = "KEY"
                           )
          )
}
\arguments{
\item{data}{Data set}

\item{log}{The log file}

\item{date_KEY}{The Unique identifier in data set. Must be same as the KEY in log file}

\item{log_column}{Column names in log file}
}
\description{
Apply cleaning log
}
\note{
It is suggested to rename the column names in the log file as below

\strong{question}: column names in the data which the changes should apply to

\strong{old_value}: the current value for the respective column

\strong{new_value}: the value to replace the old_value with

\strong{KEY}: unique identifier in log
}
