% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_relevancy_rules.R
\name{check_relevancy_rules}
\alias{check_relevancy_rules}
\title{Check Relevancy Rule}
\usage{
check_relevancy_rules(data,
                             tool_relevancy,
                             sheet_name = "data",
                             KEY = "KEY")
}
\arguments{
\item{data}{The data set for which the relevancy would be checked}

\item{tool_relevancy}{The output of create_relevancy_file function, a file in which all the relevancy formulas are converted to what is understandable by R}

\item{sheet_name}{The sheet name of data set in case it holds multiple sheets. Default value to "data"}

\item{KEY}{The Unique ID of observations. Default value to "KEY"}
}
\value{
Data frame, containing the relevancy rules broken
}
\description{
The function looks up for any value violate the relevancy rules defined in the tool.
However, the tool_relevancy parameter of the function is dependent to another function
called create_relevancy_file by which the parameter can get the suitable input
}
