% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_select_multiple.R
\name{check_select_multiple}
\alias{check_select_multiple}
\title{Check Series Cols}
\usage{
check_select_multiple(data, tool, question_separator = "_", KEY = "KEY")
}
\arguments{
\item{data}{dataframe}

\item{tool}{The survey sheet of data collection tool}

\item{question_separator}{default value is '_'}

\item{KEY}{default value is 'KEY'}
}
\description{
This function checks if the 'select_multiple' series column values (0s,1s) are based on responses in the main question.
}
