% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_url.R
\name{concat_url}
\alias{concat_url}
\title{Concatenate the image/audio url}
\usage{
concat_url(
  data,
  tool,
  server_name = "https://atrconsultingaf.surveycto.com",
  KEY = KEY,
  question_types = c("audio audit", "text audit", "audio", "image")
)
}
\arguments{
\item{data}{the dataset/dataframe}

\item{tool}{link to SCTO tool}

\item{server_name}{the server name}

\item{KEY}{the unique identifier column name - UUID}

\item{question_types}{a string vector of question types which the url should be concatenate. Defualt value c("audio audit", "text audit", "audio", "image")}
}
\description{
The \code{concat_url()} function concatenates the url for image and audio questions
}
