% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_relevancy_to_R.R
\name{convert_relevancy_to_R}
\alias{convert_relevancy_to_R}
\title{Converts the relevancy rules to R statements}
\usage{
convert_relevancy_to_R(kobo_tool)
}
\arguments{
\item{kobo_tool}{The data collection tool}
}
\value{
The list of R statements converted from Relevancy Formulas
}
\description{
Converts the relevancy rules to R statements
}
