% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_relevancy_file.R
\name{create_relevancy_file}
\alias{create_relevancy_file}
\title{Creates the relevancy file}
\usage{
create_relevancy_file(kobo_tool, ignore_reverse_check = NULL)
}
\arguments{
\item{kobo_tool}{The data collection tool's survey sheet}

\item{ignore_reverse_check}{The list of question names to be ignored for reverse checking}
}
\value{
The input for check_relevancy_rules.R function
}
\description{
#' @usage
library(atRfunctions)
create_relevancy_file(kobo_tool = kobo_tool,
ignore_reverse_check = c("Q1","Q2","Q10")
)
}
