% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_translation.R
\name{missing_translation}
\alias{missing_translation}
\title{Generate translation log}
\usage{
missing_translation(data, KEY = "KEY")
}
\arguments{
\item{data}{dataframe}

\item{KEY}{unique identifier. Default value is "KEY"}
}
\description{
#' @usage
library(atRfunctions)
missing_translation(data, KEY)
}
