% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_translation.R
\name{missing_translation}
\alias{missing_translation}
\title{Generate translation log}
\usage{
library(atRfunctions)
missing_translation(data, KEY, excluded_cols)
}
\arguments{
\item{data}{dataframe}

\item{KEY}{unique identifier; Default value is "KEY"}

\item{excluded_cols}{list of columns to be ignored}
}
\description{
Generate translation log
}
