% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_group_relevancies.R
\name{process_group_relevancies}
\alias{process_group_relevancies}
\title{Set the relevancy defined for groups on the questions inside}
\usage{
process_group_relevancies(kobo_tool)
}
\arguments{
\item{kobo_tool}{The data collection tool}
}
\value{
The list of questions and its group name involved
}
\description{
Set the relevancy defined for groups on the questions inside
}
