% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_series_cols.R
\name{update_series_cols}
\alias{update_series_cols}
\title{Update Select_multiple series columns}
\usage{
update_series_cols(data, tool, question_separator = "_")
}
\arguments{
\item{data}{dataframe}

\item{tool}{path to the XLSForm, a \code{\link[=read_xlsform]{read_xlsform()}} result list, or a
pre-read \code{survey}-sheet data frame.}

\item{question_separator}{default value is '_'}
}
\description{
This function updates the 'select_multiple' series column values (0s,1s) based on responses in the main question.
}
