% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_series_cols.R
\name{update_series_cols}
\alias{update_series_cols}
\title{Update Select_multiple series columns}
\usage{
library(atRfunctions)
update_series_cols(data,
                   tool_survey,
                   question_separator = "/")
}
\arguments{
\item{data}{dataframe}

\item{tool_survey}{The survey sheet of data collection tool}

\item{question_separator}{default value is '_'}
}
\description{
This function updates the 'select_multiple' series column values (0s,1s) based on responses in the main question.
}
