% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_series_cols.R
\name{update_series_cols}
\alias{update_series_cols}
\title{Update Select_multiple series columns}
\usage{
library(atRfunctions)
update_series_cols(data,
                   multi_vars,
                   question_separator = "/")
}
\arguments{
\item{data}{dataframe}

\item{multi_vars}{a vector of multi-select questions}

\item{question_separator}{default value is '/'}
}
\description{
This function updates the 'select_multiple' series column values (0s,1s) based on responses in the main question.
We need this because sometimes the values of select_multiple questions are updated in correction log but the series columns are not.
}
\examples{
library(atRfunctions)

dt <- data.frame(
     q1 = c("A", "A; B", "B; A", "B; C", "A; C"),
     q1_A = c(0, 0     , 0      , 0    , 0),
     q1_B = c(0, 0     , 0      , 0    , 0),
     q1_C = c(0, 0     , 0      , 0    , 0),
     q2 = c("option_A", "option_A; option_B", "option_B; option_A", "option_B; option_C", "option_A; option_C"),
     "q2_option_A" = c(1, 1     , 1      , 1    , 1),
     "q2_option_B" = c(1, 1     , 1      , 1    , 1),
     "q2_option_C" = c(1, 1     , 1      , 1    , 1)
     )

update_series_cols(dt,
                   multi_vars = c("q1", "q2"),
                   question_separator = "_")

}
